refactor(dashboard): organized dashboard layout
